# Conflicts:
#	src/main/java/com/example/demo/infra/entity/FavoriteEntity.java
#	src/main/java/com/example/demo/infra/entity/PostEntity.java
#	src/main/java/com/example/demo/infra/entity/UserEntity.java
#	src/main/java/com/example/demo/presentation/controller/pageproperty/TransitionTargetPageNameKeyword.java
#	src/main/resources/templates/fragments/header.html

+++ src/main/java/com/example/demo/domain/service/LoginService.java
@@ -28,7 +28,7 @@ public class LoginService {
 	private final UserRepository userRepository;
 	
 	    
-	public UserEntity doLogin(LoginUserForm form, Integer userType) {
+	public UserEntity doLogin(LoginUserForm form, Byte userType) {
 	    Optional<UserEntity> optionalUser =
 	            userRepository.findByEmail(form.getEmail());
 	    if (optionalUser.isEmpty()) {
